process explorer - more cleanup
