+++ src/vs/workbench/electron-sandbox/actions/developerActions.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { localize2 } from '../../../nls.js';
+import { localize, localize2 } from '../../../nls.js';
 import { INativeHostService } from '../../../platform/native/common/native.js';
 import { IEditorService } from '../../services/editor/common/editorService.js';
 import { Action2, MenuId } from '../../../platform/actions/common/actions.js';
@@ -16,6 +16,9 @@ import { KeyCode, KeyMod } from '../../../base/common/keyCodes.js';
 import { INativeWorkbenchEnvironmentService } from '../../services/environment/electron-sandbox/environmentService.js';
 import { URI } from '../../../base/common/uri.js';
 import { getActiveWindow } from '../../../base/browser/dom.js';
+import { IDialogService } from '../../../platform/dialogs/common/dialogs.js';
+import { INativeEnvironmentService } from '../../../platform/environment/common/environment.js';
+import { IProgressService, ProgressLocation } from '../../../platform/progress/common/progress.js';
 
 export class ToggleDevToolsAction extends Action2 {
 
@@ -117,5 +120,44 @@ export class ShowGPUInfoAction extends Action2 {
 	run(accessor: ServicesAccessor) {
 		const nativeHostService = accessor.get(INativeHostService);
 		nativeHostService.openGPUInfoWindow();
+	}
+}
+
+export class StopTracing extends Action2 {
+
+	static readonly ID = 'workbench.action.stopTracing';
+
+	constructor() {
+		super({
+			id: StopTracing.ID,
+			title: localize2('stopTracing', 'Stop Tracing'),
+			category: Categories.Developer,
+			f1: true
+		});
+	}
+
+	override async run(accessor: ServicesAccessor): Promise<void> {
+		const environmentService = accessor.get(INativeEnvironmentService);
+		const dialogService = accessor.get(IDialogService);
+		const nativeHostService = accessor.get(INativeHostService);
+		const progressService = accessor.get(IProgressService);
+
+		if (!environmentService.args.trace) {
+			const { confirmed } = await dialogService.confirm({
+				message: localize('stopTracing.message', "Tracing requires to launch with a '--trace' argument"),
+				primaryButton: localize({ key: 'stopTracing.button', comment: ['&& denotes a mnemonic'] }, "&&Relaunch and Enable Tracing"),
+			});
+
+			if (confirmed) {
+				return nativeHostService.relaunch({ addArgs: ['--trace'] });
+			}
+		}
+
+		await progressService.withProgress({
+			location: ProgressLocation.Dialog,
+			title: localize('stopTracing.title', "Creating trace file..."),
+			cancellable: false,
+			detail: localize('stopTracing.detail', "This can take up to one minute to complete.")
+		}, () => nativeHostService.stopTracing());
 	}
 }
